pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Pablo Pessanha <pablopessanh@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
serve = "src.scripts:server"
test = "src.scripts:tests"
manager = "src.scripts:manager(args)"


[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
djangorestframework = "^3.15.2"
django-ninja = "^1.3.0"
django-seed = "^0.3.1"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-django = "^4.9.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "src.config.settings"
python_files = ["test*.py", "*_test.py"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
